Create 2013-11-05BBBpolarity-test4b
In exploring the latest release of John McIntosh's 4D Blocks program, I wanted to explore the visualization as well as the spatial relationships among the eight poles of the cardinal directions of fourspace. Put this in your Blocks data folder and load it as a Geom to get started in the center of a configuration of pole indicators.