Update 2013-11-29 campfire1
Added supplemental columns using the formula for a circle to calculate their cornerstone placement: 

([x-a]squared)-([y-b]squared)=(r squared) 

where a and b are the coordinates of the circle center and r is the radius. 
In total, 12 cornerstones support 12 columns with two red cornerstones indicating the user's initial ana/kata orientation and blue cornerstones indicating that their location has been derived from the circle formula and not just the "six corners by four units" approach.