Create 2020-07-22RotationPaths004
R^4 contains two planes of rotation such that rotation about the origin may be computed with two angles of rotation not bound to correlate (double rotation). Colors signify paths of 2pi (360 degree) double rotation about the origin with both rotations in the same direction (isoclinic: yellow and magenta) and with rotation in opposing directions (orange and cyan). Green and red indicators are pi radians apart by a double rotation and may appear as if interior or exterior depending on observation perspective.